refactor(runner/scheduler): remove dependency index barrel
