--- packages/runner/src/scheduler/dependency-index.ts
@@ -1,4 +1,0 @@
-export * from "./dependency-graph.ts";
-export * from "./dependency-updates.ts";
-export * from "./scheduling-writes.ts";
-export * from "./trigger-index.ts";
